+++ backend/migration-sv/api/build.gradle.kts
@@ -1,0 +1,16 @@
+plugins {
+    id("convention-api")
+}
+
+group = "com.r8n.backend.migration"
+version = "0.0.1-SNAPSHOT"
+description = "user data migration service API"
+
+dependencies {
+    api(project(":messaging-api"))
+    api(project(":opinions-api"))
+    api(project(":opinions-api-integration"))
+    api(project(":users-api"))
+    api(project(":users-api-integration"))
+    api(project(":mock-api"))
+}
