+++ backend/migration-sv/service/src/main/resources/application-local.yml
@@ -1,0 +1,26 @@
+server:
+  port: ${SERVICES_MIGRATION_PORT}
+
+
+
+
+
+
+spring:
+  application.name: migration-sv
+springdoc:
+  packages-to-scan: com.r8n.backend.migration.controller
+services:
+  users:
+    url: ${INTERSERVICE_PROTOCOL}://${SERVICES_USERS_HOST}:${SERVICES_USERS_PORT}
+  opinions:
+    url: ${INTERSERVICE_PROTOCOL}://${SERVICES_OPINIONS_HOST}:${SERVICES_OPINIONS_PORT}
+  messaging:
+    url: ${INTERSERVICE_PROTOCOL}://${SERVICES_MESSAGING_HOST}:${SERVICES_MESSAGING_PORT}
+  mock:
+    url: ${INTERSERVICE_PROTOCOL}://${SERVICES_MOCK_HOST}:${SERVICES_MOCK_PORT}
+r8n:
+  security:
+    public-paths: /v3/api-docs/**, /swagger-ui.html, /swagger-ui/**, /webjars/swagger-ui/**
+    jwt:
+      public-key: ${JWT_PUBLIC_KEY}
