+++ backend/opinions-sv/api-integration/src/main/kotlin/com/r8n/backend/opinions/integration/api/OpinionListsInternalApi.kt
@@ -5,6 +5,8 @@ import com.r8n.backend.core.api.PageResponseDto
 import com.r8n.backend.opinions.api.lists.dto.OpinionListDto
 import jakarta.validation.Valid
 import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
 
 interface OpinionListsInternalApi {
     companion object {
@@ -16,4 +18,9 @@ interface OpinionListsInternalApi {
         @Valid
         pageable: PageRequestDto,
     ): PageResponseDto<OpinionListDto>
+
+    @PostMapping("/api/internal/opinion-lists/restore")
+    fun restoreOpinionList(
+        @RequestBody list: OpinionListDto,
+    )
 }
