+++ deployment/config/docker.env
@@ -11,8 +11,8 @@ SERVICES_OPINIONS_HOST=opinions
 SERVICES_OPINIONS_PORT=8080
 SERVICES_USERS_HOST=users
 SERVICES_USERS_PORT=8080
-SERVICES_EXPORT_HOST=export
-SERVICES_EXPORT_PORT=8080
+SERVICES_MIGRATION_HOST=migration
+SERVICES_MIGRATION_PORT=8080
 DATABASE_NAME=r8n
 DATABASE_USERNAME=dbuser
 DATABASE_PASSWORD=dbpassword
