+++ deployment/config/local.env
@@ -20,8 +20,8 @@ SERVICES_OPINIONS_HOST=localhost
 SERVICES_OPINIONS_PORT=8081
 SERVICES_USERS_HOST=localhost
 SERVICES_USERS_PORT=8082
-SERVICES_EXPORT_HOST=localhost
-SERVICES_EXPORT_PORT=8083
+SERVICES_MIGRATION_HOST=localhost
+SERVICES_MIGRATION_PORT=8083
 DATABASE_NAME=r8n
 DATABASE_USERNAME=dbuser
 DATABASE_PASSWORD=dbpassword
