+++ deployment/config/services.list
@@ -4,4 +4,4 @@ messaging
 opinions
 mock
 users
-export
+migration
